Average

+++ src/main/java/com/sparta/startup_be/dto/CoordinateDto.java
@@ -6,11 +6,11 @@ import lombok.Setter;
 @Getter
 @Setter
 public class CoordinateDto {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
     private Long estateid;
 
-    public CoordinateDto(float x, float y,Long estateid){
+    public CoordinateDto(double x, double y,Long estateid){
         this.x = x;
         this.y = y;
         this.estateid= estateid;
